+++ src/Module/Platform/User/Dto/UserInviteInputFactory.php
@@ -18,6 +18,7 @@ class UserInviteInputFactory implements UserInviteInputFactoryInterface
             role: Str::trimFromArray($data, 'role'),
             message: Str::trimOrNullFromArray($data, 'message'),
             disabled: (bool) ($data['disabled'] ?? false),
+            type: Str::trimOrNullFromArray($data, 'type') ?? 'backend',
         );
     }
 }